qreader/pyproject.toml
[project]
name = "qreader-service"
version = "1.0.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.18",
    "qreader>=3.16",
    "slowapi>=0.1.9",
    "pillow>=11.0.0",
    "pillow-heif>=0.18.0",
    "numpy>=2.1.0",
]

[tool.uv]
dev-dependencies = []
